Refactor appsec wordpress tests
